+++ packages/tsurlfilter/CHANGELOG.md
@@ -24,6 +24,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [AdguardBrowserExtension#2942]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2942
 
+## [3.3.4] - 2025-05-28
+
+### Changed
+
+- Updated [@adguard/agtree] to `v3.2.1`.
+
+### Fixed
+
+- Types for `NodeNext` module resolution.
+
+[3.3.4]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tsurlfilter-v3.3.4
+
 ## [3.3.3] - 2025-05-22
 
 ### Changed